requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0
piexif>=1.1.3
darkdetect>=0.8.0
tkinterdnd2>=0.3.0
